+++ CHANGES.rst
@@ -19,6 +19,7 @@ Minor Changes
 ^^^^^^^^^^^^^
 
 - fix(package): delete option likely breaking piwheels
+- fix(exc): subclass ``AttributeError`` for error raised in property
 - docs(cipher): describe a ``salt`` consistently
 - refactor(api): enforce positional or keyword only arguments